+++ src/forze/application/contracts/queue/__init__.py
@@ -1,4 +1,12 @@
+from .conformity import QueueConformity, QueueDepConformity
+from .deps import (
+    QueueReadDepKey,
+    QueueReadDepPort,
+    QueueWriteDepKey,
+    QueueWriteDepPort,
+)
 from .ports import QueueReadPort, QueueWritePort
+from .specs import QueueSpec
 from .types import QueueMessage
 
 # ----------------------- #
@@ -7,4 +15,11 @@ __all__ = [
     "QueueMessage",
     "QueueReadPort",
     "QueueWritePort",
+    "QueueConformity",
+    "QueueDepConformity",
+    "QueueSpec",
+    "QueueReadDepPort",
+    "QueueReadDepKey",
+    "QueueWriteDepPort",
+    "QueueWriteDepKey",
 ]