+++ src/forze_rabbitmq/kernel/platform/__init__.py
@@ -1,0 +1,5 @@
+from .client import RabbitMQClient, RabbitMQConfig
+
+# ----------------------- #
+
+__all__ = ["RabbitMQClient", "RabbitMQConfig"]